cleanup some names
