+++ tests/rules/Fixtures/FinalClass/SomeTrait.php
@@ -6,6 +6,8 @@
  */
 declare(strict_types=1);
 
-namespace Ibexa\Tests\PHPStan\Rules\Fixtures\FinalRule;
+namespace Ibexa\Tests\PHPStan\Rules\Fixtures\FinalClass;
 
-interface SomeInterface {}
+trait SomeTrait
+{
+}
