+++ tests/rules/Fixtures/NamingConvention/WrongName.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * @copyright Copyright (C) Ibexa AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ */
+declare(strict_types=1);
+
+namespace Ibexa\Tests\PHPStan\Rules\Fixtures\NamingConvention;
+
+interface WrongName
+{
+}
