+++ tests/rules/Fixtures/RequireInterfaceInDependencies/ConcreteClass.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @copyright Copyright (C) Ibexa AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ */
+declare(strict_types=1);
+
+namespace Ibexa\Tests\PHPStan\Rules\Fixtures\RequireInterfaceInDependencies;
+
+class ConcreteClass implements TestInterface
+{
+    public function doSomething(): void
+    {
+        // Implementation
+    }
+}
